Split shared memory discovery code into sesshm.c
